cast to correct type after value extraction from source
otherwise functions that are registered to specific types
might fail.

for example extract(day from ts) expects ts to be a long. In
    case the timestamp is in the integer range it received a
    Integer instead and failed.

+++ sql/src/test/java/io/crate/executor/transport/PagingTasksTest.java
@@ -524,7 +524,7 @@ public class PagingTasksTest extends BaseTransportExecutorTest {
         int numRows = randomIntBetween(1, 4096);
         Object[][] bulkArgs = new Object[numRows][];
         for (int l = 0; l < numRows; l++) {
-            bulkArgs[l] = new Object[]{l};
+            bulkArgs[l] = new Object[]{ (long)l};
         }
         sqlExecutor.exec("insert into ids values (?)", bulkArgs);
         sqlExecutor.refresh("ids");
